requirements.txt
alabaster==0.7.12
Babel==2.6.0
certifi==2019.3.9
chardet==3.0.4
docutils==0.14
humanfriendly==4.18
idna==2.8
imagesize==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
nose==1.3.7
numpy==1.16.3
packaging==19.0
pandas==0.24.2
py4j==0.10.7
pyarrow==0.13.0
Pygments==2.3.1
PyMySQL==0.9.3
pyparsing==2.4.0
pyspark==2.4.2
python-dateutil==2.8.0
python-dotenv==0.10.1
pytz==2019.1
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
sklearn==0.0
snowballstemmer==1.2.1
Sphinx==2.0.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SQLAlchemy==1.3.5
urllib3==1.24.2
